+++ log-viewer/modules/components/LogTitle.ts
@@ -5,8 +5,8 @@ import { provideVSCodeDesignSystem, vsCodeButton } from '@vscode/webview-ui-tool
 import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { globalStyles } from '../global.styles.js';
 import { hostService } from '../services/VSCodeService.js';
+import { globalStyles } from '../styles/global.styles.js';
 import { skeletonStyles } from './skeleton/skeleton.styles.js';
 
 provideVSCodeDesignSystem().register(vsCodeButton());
