+++ log-viewer/modules/components/app-header/AppHeader.ts
@@ -10,16 +10,16 @@ import {
 import { LitElement, css, html, unsafeCSS } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 
+import codiconStyles from '../../styles/codicon.css';
+import { globalStyles } from '../../styles/global.styles.js';
+import { RootNode } from '../../timeline/Timeline.js';
+import '../../timeline/TimelineView.js';
 import '../analysis-view/AnalysisView.js';
 import '../calltree-view/CalltreeView';
 import '../database-view/DatabaseView.js';
-import { globalStyles } from '../global.styles.js';
 import '../log-levels/LogLevels.js';
 import '../navbar/NavBar.js';
 import { Notification } from '../notifications/NotificationPanel.js';
-import codiconStyles from '../styles/codicon.css';
-import { RootNode } from '../timeline/Timeline.js';
-import '../timeline/TimelineView.js';
 
 provideVSCodeDesignSystem().register(vsCodePanelTab(), vsCodePanelView(), vsCodePanels());
