+++ log-viewer/modules/components/calltree-view/TreeView.scss
@@ -1,4 +1,4 @@
-@import '../datagrid/style/DataGrid.scss';
+@import '../../datagrid/style/DataGrid.scss';
 
 #call-tree-container {
   display: flex;
