+++ log-viewer/modules/components/database-view/DatabaseView.scss
@@ -1,4 +1,4 @@
-@import '../datagrid/style/DataGrid.scss';
+@import '../../datagrid/style/DataGrid.scss';
 
 #db-container {
   overflow: scroll;
