+++ log-viewer/modules/components/log-levels/LogLevels.ts
@@ -4,9 +4,9 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
 
-import { skeletonStyles } from '../components/skeleton/skeleton.styles.js';
-import { globalStyles } from '../global.styles.js';
-import { LogSetting } from '../parsers/TreeParser.js';
+import { LogSetting } from '../../parsers/TreeParser.js';
+import { globalStyles } from '../../styles/global.styles.js';
+import { skeletonStyles } from '../skeleton/skeleton.styles.js';
 
 @customElement('log-levels')
 export class LogLevels extends LitElement {
