+++ log-viewer/modules/components/log-viewer/LogViewer.ts
@@ -4,9 +4,6 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import './app-header/AppHeader.js';
-import { globalStyles } from './global.styles.js';
-import { Notification } from './notifications/NotificationPanel.js';
 import parseLog, {
   RootNode,
   TruncationColor,
@@ -15,8 +12,11 @@ import parseLog, {
   totalDuration,
   truncateLog,
   truncated,
-} from './parsers/TreeParser.js';
-import { hostService } from './services/VSCodeService.js';
+} from '../../parsers/TreeParser.js';
+import { hostService } from '../../services/VSCodeService.js';
+import { globalStyles } from '../../styles/global.styles.js';
+import '../app-header/AppHeader.js';
+import { Notification } from '../notifications/NotificationPanel.js';
 
 @customElement('log-viewer')
 export class LogViewer extends LitElement {
