+++ log-viewer/modules/components/notifications/NotificationTag.ts
@@ -5,8 +5,8 @@ import { provideVSCodeDesignSystem, vsCodeButton } from '@vscode/webview-ui-tool
 import { LitElement, css, html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 
-import '../components/BadgeBase.js';
-import { globalStyles } from '../global.styles.js';
+import { globalStyles } from '../../styles/global.styles.js';
+import '../BadgeBase.js';
 import './NotificationPanel.js';
 
 provideVSCodeDesignSystem().register(vsCodeButton());
