+++ log-viewer/modules/soql/SOQLLinter.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Certinia Inc. All rights reserved.
  */
 import { type Stack } from '../Database.js';
-import { SOQLParser, SOQLTree } from '../parsers/SOQLParser.js';
+import { SOQLParser, SOQLTree } from './SOQLParser.js';
 
 //todo : need a general concept of stack (to pull out linter)
 export class SOQLLinter {
