+++ log-viewer/modules/styles/global.styles.ts
