+++ log-viewer/modules/styles/notification.styles.ts
