+++ log-viewer/modules/timeline/Timeline.ts
@@ -3,7 +3,7 @@
  */
 //TODO:Refactor - usage should look more like `new TimeLine(timelineContainer, {tooltip:true}:Config)`;
 import formatDuration, { debounce } from '../Util.js';
-import { goToRow } from '../calltree-view/CalltreeView.js';
+import { goToRow } from '../components/calltree-view/CalltreeView.js';
 import { Method, RootNode, TimedNode, type TimelineKey, truncated } from '../parsers/TreeParser.js';
 
 export { RootNode };
